backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
# Emergency calling dependencies
requests==2.31.0
pytz==2023.3